+++ src/translations/en/ckeditor.php
@@ -36,6 +36,7 @@ return [
     'Links' => 'Links',
     'No transform' => 'No transform',
     'Purify HTML' => 'Purify HTML',
+    'Raw content only' => 'Raw content only',
     'Removes any potentially-malicious code on save, by running the submitted data through {link}.' => 'Removes any potentially-malicious code on save, by running the submitted data through {link}.',
     'Show word count' => 'Show word count',
     'Site: {name}' => 'Site: {name}',